[PyTorch] Support borrowing in/out Tensors in TensorIterator
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55690

Just change `OperandInfo::tensor` and
`TensorIteratorConfig::tensors` to hold `c10::MaybeOwned<Tensor>`, and
deal with the consequent pointer syntax. Had to C10_ALWAYS_INLINE
OperandInfo to preserve existing inlining behavior for whatever
compiler-idiosyncratic reason.

This is a separate diff from usage to enable measuring the cost of
support, and because there is no reason not to send it separately.

We probably should not land this without a plan to migrate a lot of
TensorIterator use cases to use either borrowing or structured kernels
& borrowing.
ghstack-source-id: 126590815

Differential Revision: [D27607293](https://our.internmc.facebook.com/intern/diff/D27607293/)